Make interactive mode optional
Add a configuration option for interactive usage. When `interactive` is
enabled (either in the configuration file or via the `--interactive/-i`
flag), the user can be prompted to make decisions. The only supported
feature at the moment is selecting a TV series from a list of multiple
search results - in the new `nielsen.tv.select_series` function. When
disabled, a different remote API is used to query the series title which
has a higher "fuzziness" value and always returns a single result.

Disable `interactive` in the default `CONFIG` variable, but enable it in
the example `nielsen.ini`.

Bump the version to `2.0.0`.

Resolve #64.