Use pathlib for handling file paths and manipulation
Add support for `pathlib` and use it to handle generating file paths
and extracting information from them rather than `os.path`.

Add a `nielsen.files` module which handles most of the underlying file
operations (making calls out to `shutil` for things like `chown` and
`move`). This abstraction allows not only for more reusable code, but
smaller functions to test and removes the overhead of a lot of context
for the API level functions (most notably `organize_file`).

Favor returning `Path` objects from API functions which modify files in
some way. Returning the most up-to-date version of the `Path` object
allows the API to be more composable and extendable.

Resolves #73.