Use pathlib for handling file paths and manipulation (#74)
Add support for `pathlib` and use it to handle generating file paths
and extracting information from them rather than `os.path`.

Add a `nielsen.files` module which handles most of the underlying file
operations (making calls out to `shutil` for things like `chown` and
`move`). This abstraction allows not only for more reusable code, but
smaller functions to test and removes the overhead of a lot of context
for the API level functions (most notably `organize_file`).

Favor returning `Path` objects from API functions that modify files in
some way. Returning the most up-to-date version of the `Path` object
allows the API to be more composable and extendable.

Use an intermediate value for defining the filename format for processed
files. This opens up possibilities for customization.

Bump version to `2.1.0`.

Resolves #71.

Resolves #73.