Rename `filter_filename` to `sanitize_filename`
Rename 'filter_filename` to `sanitize_filename` and add a `system_type`
parameter to change which OS type it will sanitize the filename for. By
default, this is set to `os.name` and supports `posix` and `nt`.

Streamline the behavior of `filter_series`. This function now returns
the given string in title case unless a custom name is found.

Update list of supported Python versions to a minimum of Python 3.6
(which is required for f-strings).

Bump version to `2.1.3`.