Add the ability to customize output filenames (#80)
Add a new option to the configuration: `format`. This option determines
the new filename of files processed by `nielsen`. It supports the
following tokens: `{series}`, `{season}`, `{episode}`, `{title}`,
`{extension}`.

The default format remains unchanged and has been added to the sample
`nielsen.ini` to preserve existing behavior and suggest the ability to
modify it.

Resolves #21.