Restore dry-run functionality
Add a `dryrun` parameter to the `organize_file` and `process_file`
functions in the `api` module. Get the value for this parameter from the
`CONFIG` object by default.

Add checks for the `dryrun` parameter before performing operations which
modify files or the filesystem. This functionality was broken when
adding support for `pathlib`.

Add unit test for `organize_file`.

Bump version to 2.2.2.

Resolve #83.