Refactor Media class to more easily implement subclasses
Change `nielsen.media.Media` to a normal class to allow for easier
testing of the functions that subclasses inherit from it.

Add tests for the `Media` base class now that objects of that type can
be instantiated.

Add `section` and `library` properties to the `Media` base class to make
it easier to get configuration options from the correct section and
ensure the destination directory for organizing files is always an
absolute path to avoid ambiguity.

Raise a `NotImplementedError` when trying to call `Media.infer()` to
force sub-classes to implement it.

Set default configuration operations with a `DEFAULT` dictionary at the
module level to avoid forcing a call to `nielsen.config.load_config()`
to set the defaults. If the module is imported, the defaults should be
set.

Add a brief description to the default configuration options.

Use full option names in `vim` modelines for clarity.