Add `metadata` property to `Media` base class
Add a `metadata` property to the base `Media` class. The getter and
setter for this property call a `_get_metadata` and `_set_metadata`
method, respectively. These methods are not implemented in the base
class and will raise a `NotImplementedError`, but each subclass can
implement them without having to worry about the reference being masked
by inheritance.

Add a `transform` method to the base `Media` class which takes a field
as an argument and attempts to replace the given value with one found in
the corresponding config section. This should make normalizing series
names much easier (transforming "Marvel's Agents of S.H.I.E.L.D." to
"Agents of SHIELD" or vice versa).