Add `nielsen.fetcher` module and a `Fetcher` `Protocol` class (#95)
Add `nielsen.fetcher.TVMaze`, which is able to make requests to the
`TVMaze` API to fetch additional metadata about a given `Media` object.

Add tests for `nielsen.fetcher`.

Add `bin/pickler.py`, which makes actual requests to external APIs, and
uses Python's `pickle` module to save those responses to be used as test
fixtures.

Add a `metadata` property to the base `Media` class. These methods are not
implemented in the base class and will raise a `NotImplementedError`.

Add a `transform` method to the base `Media` class which takes a field as an
argument and attempts to replace the given value with one found in the
corresponding config section. This should make normalizing series names much
easier (e.g. transforming "Marvel's Agents of S.H.I.E.L.D." to "Agents of
SHIELD" or vice versa).

Add `nielsen.config.update_config` function to update a config file with
new sections/options.

Add `requests` package as a dependency.

Switch to `poetry` as the virtual environment, package, and dependency
management tool of choice.

Resolves #90, resolves #92.