Make `Media.path` required (#102)
Make `Media.path` a required property of all `Media` objects. Set the
`path` in `Media.__post_init__` to ensure it goes through the setter,
which coerces the value to a `pathlib.Path` or raises a `TypeError`.
This makes it more difficult to create dummy media objects that don't
represent a file on disk, but there's no obvious use-case that requires
such an object (besides convenience in creating objects for unit tests).

Ensure that `Media.path` is a file (even if it doesn't exist) when
setting the property.

Ensure that `Media.library` is a directory (even if it doesn't exist)
when setting the property.

Add `types-requests` as a development dependency for better type checking.

Omit `nielsen/logging.py` from `coverage` tests and reports.

Resolve #101.