Reorder the first two TV matching patterns (#111)
Reorder the first two patterns in `nielsen.media.TV.patterns`. Filenames
with a lot of 4 digit numbers in them (series year, a year in the title,
and a `1080p` resolution) were matching the first pattern with the year
in the title mistaken for the series year and the resolution tag being
interpreted as the season and episode numbers. Swapping the order of the
first two patterns matches the less ambiguous `SxxExx` format first,
which should prevent similar types of issues.

Add the problematic filename as a test case.

Resolve #110.